Made the user-facing information for IDs, indexes, and numbers consistent.

+++ source/source/game_state/gui_editor/editor.cpp
@@ -793,7 +793,7 @@ void GuiEditor::setSelectionStatusText() {
         if(itemSelection.hasAny()) {
             setStatus(
                 "Selected " +
-                amountStr(itemSelection.getCount(), "item") + "."
+                amountStr((int) itemSelection.getCount(), "item") + "."
             );
         }
         break;
